// types.generated.ts
export type Maybe<T> = T;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  AWSDate: string;
  AWSDateTime: string;
  AWSEmail: string;
  AWSIPAddress: string;
  AWSJSON: string;
  AWSPhone: string;
  AWSTime: string;
  AWSTimestamp: string;
  AWSURL: string;
};

export type AvailabilityStats = {
  __typename?: 'AvailabilityStats';
  month: Maybe<Scalars['String']>;
  slots: Maybe<Scalars['Int']>;
};

export type Company = Node & {
  __typename?: 'Company';
  comments: Maybe<Scalars['String']>;
  createdAt: Scalars['AWSDateTime'];
  endsOn: Maybe<Scalars['AWSDateTime']>;
  id: Scalars['ID'];
  name: Maybe<Scalars['String']>;
  nodeId: Scalars['ID'];
};

export type CompanyId = {
  __typename?: 'CompanyId';
  id: Scalars['ID'];
};

export type CompanyInput = {
  comments?: Maybe<Scalars['String']>;
  endsOn?: Maybe<Scalars['AWSDateTime']>;
  id?: Maybe<Scalars['ID']>;
  name: Scalars['String'];
};

export type EventInfo = {
  __typename?: 'EventInfo';
  created: Maybe<Scalars['AWSDateTime']>;
  htmlLink: Maybe<Scalars['String']>;
  status: Maybe<Scalars['String']>;
  updated: Maybe<Scalars['AWSDateTime']>;
};

export type Focuscheck = Node & {
  __typename?: 'Focuscheck';
  createdAt: Maybe<Scalars['AWSDateTime']>;
  id: Scalars['ID'];
  intentionForMe: Maybe<Scalars['String']>;
  intentionForOthers: Maybe<Scalars['String']>;
  investigation: Maybe<Scalars['String']>;
  nodeId: Scalars['ID'];
  participantId: Maybe<Scalars['ID']>;
  sessionId: Maybe<Scalars['ID']>;
  successCriteria: Maybe<Scalars['String']>;
  title: Maybe<Scalars['String']>;
  updatedAt: Maybe<Scalars['AWSDateTime']>;
};

export type Group = Node & {
  __typename?: 'Group';
  comments: Maybe<Scalars['String']>;
  createdAt: Scalars['AWSDateTime'];
  deletedAt: Maybe<Scalars['AWSDateTime']>;
  id: Scalars['ID'];
  name: Scalars['String'];
  nodeId: Scalars['ID'];
  participants: Array<Participant>;
  trainer: Trainer;
  trainerId: Scalars['ID'];
  updatedAt: Maybe<Scalars['AWSDateTime']>;
};

export type GroupId = {
  __typename?: 'GroupId';
  id: Scalars['ID'];
};

export type GroupInput = {
  comments?: Maybe<Scalars['String']>;
  id?: Maybe<Scalars['ID']>;
  name: Scalars['String'];
  trainerId: Scalars['ID'];
};

export type GroupSession = Node & {
  __typename?: 'GroupSession';
  comments: Maybe<Scalars['String']>;
  createdAt: Scalars['AWSDateTime'];
  deletedAt: Maybe<Scalars['AWSDateTime']>;
  duration: Scalars['Int'];
  group: Group;
  groupId: Scalars['ID'];
  id: Scalars['ID'];
  nodeId: Scalars['ID'];
  start: Scalars['AWSDateTime'];
  updatedAt: Maybe<Scalars['AWSDateTime']>;
};
// ...
export type Message = Node & {
  __typename?: 'Message';
  createdAt: Scalars['AWSDateTime'];
  fromId: Scalars['ID'];
  id: Scalars['ID'];
  message: Scalars['String'];
  nodeId: Scalars['ID'];
  read: Maybe<Scalars['Boolean']>;
  sessionId: Maybe<Scalars['ID']>;
  toId: Scalars['ID'];
};
// ...
export type Node = {
  nodeId: Scalars['ID'];
};

export type Note = Node & {
  __typename?: 'Note';
  createdAt: Maybe<Scalars['AWSDateTime']>;
  id: Scalars['ID'];
  nodeId: Scalars['ID'];
  note: Maybe<Scalars['String']>;
  participantId: Maybe<Scalars['ID']>;
  privacy: Maybe<Scalars['Int']>;
  sessionId: Maybe<Scalars['ID']>;
  trainerId: Maybe<Scalars['ID']>;
  updatedAt: Maybe<Scalars['AWSDateTime']>;
};

export type NoteId = {
  __typename?: 'NoteId';
  id: Scalars['ID'];
};

export type Participant = Node & {
  __typename?: 'Participant';
  company: Maybe<Company>;
  companyId: Maybe<Scalars['ID']>;
  createdAt: Scalars['AWSDateTime'];
  focusChecks: Maybe<Array<Focuscheck>>;
  group: Maybe<Group>;
  groupId: Maybe<Scalars['ID']>;
  groupSessions: Array<GroupSession>;
  language: Maybe<Language>;
  nodeId: Scalars['ID'];
  note: Maybe<Scalars['String']>;
  notes: Maybe<Array<Note>>;
  phoneOS: Maybe<PhoneOs>;
  possibilities: Maybe<Scalars['String']>;
  sessionStats: SessionStats;
  sessions: Array<Session>;
  sessionsLeft: Maybe<Scalars['Int']>;
  status: ParticipantStatus;
  statusQuo: Maybe<Scalars['String']>;
  statusQuos: Array<StatusQuo>;
  subscriptionLength: Maybe<Scalars['Int']>;
  tasks: Maybe<Array<Task>>;
  tel: Maybe<Scalars['String']>;
  trainer: Trainer;
  trainerId: Maybe<Scalars['ID']>;
  traitName: Maybe<Scalars['String']>;
  type: Maybe<ParticipantType>;
  unenrolledAt: Maybe<Scalars['AWSDateTime']>;
  updatedAt: Maybe<Scalars['AWSDateTime']>;
  user: User;
  userId: Scalars['ID'];
  vision: Maybe<Array<Maybe<Scalars['String']>>>;
};
// ...
export type PublicParticipant = Node & {
  __typename?: 'PublicParticipant';
  nodeId: Scalars['ID'];
  sessions: Array<PublicSession>;
  user: PublicUser;
  userId: Scalars['ID'];
};

export type PublicSession = Node & {
  __typename?: 'PublicSession';
  nodeId: Scalars['ID'];
  start: Scalars['AWSDateTime'];
};

export type PublicUser = Node & {
  __typename?: 'PublicUser';
  name: Scalars['String'];
  nodeId: Scalars['ID'];
};

export type Query = {
  __typename?: 'Query';
  allSessionsDatesAsSelf: Array<SessionDate>;
  companyGet: Maybe<Company>;
  companyList: Array<Company>;
  focuscheckGetAsParticipant: Maybe<Focuscheck>;
  focuscheckListAsParticipant: Maybe<Array<Maybe<Focuscheck>>>;
  getSignedUrl: Maybe<SignedUrl>;
  groupGet: Maybe<Group>;
  groupList: Array<Group>;
  groupSessionGet: Maybe<GroupSession>;
  groupSessionList: Array<GroupSession>;
  groupSessionListAsSelf: Array<GroupSession>;
  messageList: Maybe<Array<Maybe<Message>>>;
  messageListAsTrainer: Maybe<Array<Maybe<Message>>>;
  noteAsParticipant: Maybe<Note>;
  noteListAsParticipant: Maybe<Array<Maybe<Note>>>;
  noteListAsTrainer: Maybe<Array<Maybe<Note>>>;
  participantGet: Maybe<Participant>;
  participantGetAsParticipant: Maybe<Participant>;
  participantGetAsTrainer: Maybe<Participant>;
  participantGetPublic: Maybe<PublicParticipant>;
  participantList: Array<Participant>;
  participantsListForGroup: Array<Participant>;
  sessionGet: Maybe<Session>;
  sessionGetAsAdmin: Maybe<Session>;
  sessionGetAsTrainer: Maybe<Session>;
  sessionList: Array<Session>;
  sessionListAsParticipant: Array<Session>;
  sessionListAsSelf: Array<Session>;
  taskAsParticipant: Task;
  taskListAsParticipant: Maybe<Array<Task>>;
  trainerAvailability: Maybe<Array<Maybe<TrainerAvailability>>>;
  trainerAvailabilityAsAdmin: Maybe<Array<Maybe<TrainerAvailability>>>;
  trainerAvailabilityAsTrainer: Maybe<Array<Maybe<TrainerAvailability>>>;
  trainerCalendars: Maybe<Array<Maybe<TrainerCalendars>>>;
  trainerGet: Maybe<Trainer>;
  trainerGetSelf: Maybe<Trainer>;
  trainerList: Maybe<Array<Maybe<Trainer>>>;
  trainersAvailability: Maybe<Array<Maybe<TrainersAvailability>>>;
};
// ...
export type Session = Node & {
  __typename?: 'Session';
  createdAt: Maybe<Scalars['AWSDateTime']>;
  deletedAt: Maybe<Scalars['AWSDateTime']>;
  duration: Maybe<Scalars['Int']>;
  eventId: Maybe<Scalars['ID']>;
  eventInfo: Maybe<EventInfo>;
  id: Scalars['ID'];
  markedRecognitionAt: Maybe<Scalars['AWSDateTime']>;
  markedStatusQuoAt: Maybe<Scalars['AWSDateTime']>;
  nodeId: Scalars['ID'];
  notes: Array<Note>;
  participant: Maybe<Participant>;
  participantId: Maybe<Scalars['ID']>;
  sessionNo: Maybe<Scalars['Int']>;
  start: Scalars['AWSDateTime'];
  status: Maybe<SessionStatus>;
  trainer: Maybe<Trainer>;
  trainerId: Maybe<Scalars['ID']>;
  updatedAt: Maybe<Scalars['AWSDateTime']>;
};

export type SessionBooked = {
  __typename?: 'SessionBooked';
  id: Scalars['ID'];
};

export type SessionDate = {
  __typename?: 'SessionDate';
  start: Scalars['AWSDateTime'];
  type: Scalars['String'];
};

export type SessionStats = {
  __typename?: 'SessionStats';
  booked: Scalars['Int'];
  held: Scalars['Int'];
  next: Maybe<Scalars['AWSDateTime']>;
  noshow: Scalars['Int'];
  previous: Maybe<Scalars['AWSDateTime']>;
  total: Scalars['Int'];
  upcoming: Scalars['Int'];
};

export enum SessionStatus {
  Complete = 'complete',
  Deleted = 'deleted',
  Noshow = 'noshow',
}

export type SignedUrl = {
  __typename?: 'SignedUrl';
  formdata: Scalars['AWSJSON'];
};
// ...
export type Task = Node & {
  __typename?: 'Task';
  completedAt: Maybe<Scalars['AWSDateTime']>;
  createdAt: Scalars['AWSDateTime'];
  description: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  nodeId: Scalars['ID'];
  participantId: Maybe<Scalars['ID']>;
  repeatType: Maybe<RepeatType>;
  trainerId: Maybe<Scalars['ID']>;
};

export type Trainer = Node & {
  __typename?: 'Trainer';
  airtable: Maybe<Scalars['String']>;
  background: Maybe<Scalars['String']>;
  calendars: Maybe<Array<Maybe<Scalars['String']>>>;
  createdAt: Scalars['AWSDateTime'];
  facetimeId: Maybe<Scalars['String']>;
  keywords: Maybe<Array<Maybe<Scalars['String']>>>;
  nodeId: Scalars['ID'];
  participants: Maybe<Array<Participant>>;
  skypeId: Maybe<Scalars['String']>;
  tel: Maybe<Scalars['String']>;
  updatedAt: Maybe<Scalars['AWSDateTime']>;
  user: User;
  userId: Scalars['ID'];
};

export type TrainerAvailability = {
  __typename?: 'TrainerAvailability';
  end: Maybe<Scalars['AWSDateTime']>;
  start: Maybe<Scalars['AWSDateTime']>;
};

export type TrainerCalendars = {
  __typename?: 'TrainerCalendars';
  id: Scalars['String'];
  summary: Scalars['String'];
};
// ...
export type TrainersAvailability = {
  __typename?: 'TrainersAvailability';
  availability: Maybe<Array<Maybe<AvailabilityStats>>>;
  trainerId: Scalars['ID'];
};

export type User = Node & {
  __typename?: 'User';
  avatar: Maybe<Scalars['AWSURL']>;
  email: Scalars['AWSEmail'];
  id: Scalars['ID'];
  name: Scalars['String'];
  nodeId: Scalars['ID'];
};
// ...
export type SessionsBookScreenQuery = {
  __typename?: 'Query';
  sessionListAsParticipant: Array<{
    __typename?: 'Session';
    nodeId: string;
    eventId: Maybe<string>;
    start: string;
    duration: Maybe<number>;
    status: Maybe<SessionStatus>;
    sessionNo: Maybe<number>;
    trainerId: Maybe<string>;
    trainer: Maybe<{
      __typename?: 'Trainer';
      facetimeId: Maybe<string>;
      skypeId: Maybe<string>;
      tel: Maybe<string>;
      user: { __typename?: 'User'; name: string; avatar: Maybe<string> };
    }>;
    notes: Array<{ __typename?: 'Note'; note: Maybe<string>; nodeId: string }>;
  }>;
};
// ...
export const SessionContentFragmentDoc = gql`
  fragment SessionContent on Session {
    nodeId
    eventId
    start
    duration
    status
    sessionNo
    trainerId
    trainer {
      facetimeId
      skypeId
      tel
      user {
        name
        avatar
      }
    }
    notes {
      note
      nodeId
    }
  }
`;
// ...
export const SessionsBookScreenDocument = gql`
  query sessionsBookScreen {
    sessionListAsParticipant {
      ...SessionContent
    }
  }
  ${SessionContentFragmentDoc}
`;